spell removal and spell/spell book api refactor
when in a spell book view the add spell button becomes a remove spell button. The spell updating api changes make for a cleaner interface and now update state by making a copy of the current state, updating the copy and then updating. This has removed the bug where the localStorage lags behind the state by one change.